requirements.txt
behave             == 1.2.5
docker-py          == 1.2.3
docopt             == 0.6.2
fn                 == 0.4.3
nose               == 1.3.7
pyyaml             == 3.11
rednose            == 0.4.3
requests[security] == 2.7.0
scripttest         == 1.3
twine              == 1.5.0
dockerpty          == 0.3.2
pexpect            == 3.3
